Fixed bugs in pose estimation and visualization, implemented tracking pipeline

+++ include/BundleAdjuster.h
@@ -69,5 +69,5 @@ struct ReprojectionError3D
 class BundleAdjuster {
 public:
         BundleAdjuster();
-        Sophus::SE3d optimizeLocalPoseBA_ceres(std::vector<cv::Point3d> p3d,std::vector<cv::Point2d> p2d,Eigen::Matrix3d K,Sophus::SE3d pose);
+        Sophus::SE3d optimizeLocalPoseBA_ceres(std::vector<cv::Point3f> p3d,std::vector<cv::Point2f> p2d,Eigen::Matrix3d K,Sophus::SE3d pose);
 };
